refactor: move mapping to util package
