+++ app/src/main/java/aeropresscipe/divinelink/aeropress/savedrecipes/SavedRecipeDomain.kt
@@ -1,7 +1,7 @@
 package aeropresscipe.divinelink.aeropress.savedrecipes
 
 import aeropresscipe.divinelink.aeropress.generaterecipe.models.Recipe
-import aeropresscipe.divinelink.aeropress.mapping.MappingModel
+import aeropresscipe.divinelink.aeropress.util.mapping.MappingModel
 import androidx.room.Entity
 import androidx.room.PrimaryKey
