+++ app/src/main/java/aeropresscipe/divinelink/aeropress/util/mapping/Factory.kt
@@ -1,4 +1,4 @@
-package aeropresscipe.divinelink.aeropress.mapping
+package aeropresscipe.divinelink.aeropress.util.mapping
 
 import android.view.ViewGroup
