+++ app/src/main/java/aeropresscipe/divinelink/aeropress/util/mapping/MappingModel.kt
@@ -1,4 +1,4 @@
-package aeropresscipe.divinelink.aeropress.mapping
+package aeropresscipe.divinelink.aeropress.util.mapping
 
 interface MappingModel {
     fun areItemsTheSame(newItem: Any): Boolean
