+++ app/src/main/java/aeropresscipe/divinelink/aeropress/util/mapping/MappingViewHolder.kt
@@ -1,4 +1,4 @@
-package aeropresscipe.divinelink.aeropress.mapping
+package aeropresscipe.divinelink.aeropress.util.mapping
 
 import android.content.Context
 import android.view.View
